requirements.txt
beautifulsoup4
boto3
botocore
certifi
cffi
charset-normalizer
cryptography
et-xmlfile
idna
jmespath
msal
mysql-connector
numpy
Office365-REST-Python-Client
openpyxl
pandas
pdfminer.six
pdfplumber
pillow
pycparser
PyJWT
PyPDF2
pypdfium2
python-dateutil
python-dotenv
pytz
requests
s3transfer
six
soupsieve
typing_extensions
tzdata
urllib3
XlsxWriter
